+++ user/src/database/repository/seller.js
@@ -11,11 +11,16 @@ const {
 } = require("../../utils/app-error");
 
 class SellerRepository {
-  async CreateSeller({userId, company, logo, badgeId, description, addressId}) {
+  async CreateSeller({userId, company, logo, badgeId, description}) {
     try {
       const existingUser = await userModel.findById(userId);
       const phoneNo = existingUser.phoneNo;
       const email = existingUser.email;
+
+      const existingSeller = await sellerModel.findOne({userId: userId});
+      if(existingSeller){
+        return false;
+      }
 
       const { path: imagePath, mimetype: mimeType, size: imageSize } = logo;
 
@@ -26,7 +31,6 @@ class SellerRepository {
         badgeId: badgeId,
         description: description,
         contact: [{ phoneNo: phoneNo, email: email }],
-        Address: addressId,
       });
       const sellerResult = await seller.save();
 
@@ -34,7 +38,7 @@ class SellerRepository {
       if (history) {
         history.log.push({
           objectId: sellerResult._id,
-          action: "profile created",
+          action: "Seller profile created",
           date: new Date().toISOString(),
           time: Date.now(),
         });
@@ -43,21 +47,33 @@ class SellerRepository {
 
       return sellerResult;
     } catch (error) {
-      throw new APIError("API Error", STATUS_CODES.INTERNAL_ERROR, "Error on Create Address");
+      throw new APIError("API Error", STATUS_CODES.INTERNAL_ERROR, error.message);
+    }
+  }
+
+  async GetProfile ({userId}) {
+    try{
+      const sellerProfile = await sellerModel.find({ userId:userId });
+      if(!sellerProfile){
+        return null;
+      }
+      return sellerProfile;
+    } catch (error) {
+      throw new APIError("API Error", STATUS_CODES.INTERNAL_ERROR, error.message);
     }
   }
 
   async updateSeller({ userId, data, logo }){
     try {
-      const { datas } = data;
+      const datas = data;
       const { path: imagePath, mimetype: mimeType, size: imageSize } = logo;
 
-      const updatedSeller = await Seller.findOneAndUpdate(
+      const updatedSeller = await sellerModel.findOneAndUpdate(
         { userId: userId },
         {
           datas,
           $set: {
-            'logo.data': imagePath,
+            'logo.imagePath': imagePath,
             'logo.mimeType': mimeType,
             'logo.imageSize': imageSize,
           }
@@ -65,12 +81,12 @@ class SellerRepository {
         { new: true }
       );
       if (!updatedSeller){
-        throw new APIError('API Error', STATUS_CODES.NOT_FOUND, 'seller profile not found');
+        return false;
       }
 
       return updatedSeller;
     } catch (error) {
-      throw new APIError("API Error", STATUS_CODES.INTERNAL_ERROR, "Error on Create Address");
+      throw new APIError("API Error", STATUS_CODES.INTERNAL_ERROR, error.message);
     }
   }
 
@@ -80,6 +96,7 @@ class SellerRepository {
 
       if (user) {
         const Address = new addressModel({
+          userId: userId,
           address1: address1,
           address2: address2,
           city: city,
@@ -88,12 +105,6 @@ class SellerRepository {
           country: country,
         });
         const addressResult = await Address.save();
-
-        const updateSeller = await sellerModel.findOneAndUpdate(
-          { userId },
-          { $push: { Address: addressResult._id } },
-          { new: true }
-        );
 
         const history = await historyModel.findOne({ userId });
         if (history) {
@@ -108,22 +119,44 @@ class SellerRepository {
 
         return addressResult;
       }
-    } catch (err) {
-      throw new APIError("API Error", STATUS_CODES.INTERNAL_ERROR, "Error on Create Address");
+    } catch (error) {
+      throw new APIError("API Error", STATUS_CODES.INTERNAL_ERROR, error.message);
     }
   }
 
-  async DeleteUser(userId){
-    try{
-      const user = await userModel.findByIdAndDelete(userId);
-      const profile = await sellerModel.findOneAndDelete({ userId });
-      for (const addr of profile.Address) {
-        const addressData = await addressModel.findByIdAndDelete(addr);
+  async DeleteAddress({ userId, addressId }){
+    try {
+      const address = await addressModel.findOneAndDelete({userId: userId, _id: addressId});
+      if(!address){
+        return false;
       }
-      const password = await passwordModel.findOneAndDelete({ userId });
-      const history = await historyModel.findOneAndDelete({ userId });
-      if(!user || !profile || !password || !history){
-        throw new APIError('Deletion Failed', STATUS_CODES.BAD_REQUEST, 'Failed to delete the User');
+      return true;
+    } catch (error) {
+      throw new APIError("API Error", STATUS_CODES.INTERNAL_ERROR, error.message);
+    }
+  }
+
+  async GetAddress({userId}) {
+    try{
+      const addresses = await addressModel.find({ userId:userId }).sort('-createdAt');
+      if(!addresses){
+        return null;
+      }
+      return addresses;
+    } catch (error) {
+      throw new APIError("API Error", STATUS_CODES.INTERNAL_ERROR, error.message);
+    }
+  }
+
+  async DeleteUser({userId}){
+    try{
+      const user = await userModel.findOneAndDelete({_id: userId});
+      const profile = await sellerModel.findOneAndDelete({ userId: userId });
+      const address = await addressModel.deleteMany({ userId: userId });
+      const password = await passwordModel.findOneAndDelete({ userId: userId });
+      const history = await historyModel.findOneAndDelete({ userId: userId });
+      if(!user || !profile || !address || !password || !history){
+        return false;
       }
       return true;
     } catch(error) {
